Update fsview to only check xattr on linux
