+++ internal/fsview/overlay.go
@@ -20,13 +20,8 @@ import (
 	"errors"
 	"io"
 	"io/fs"
-	"os"
 	"path"
 	"sort"
-	"syscall"
-
-	"github.com/erofs/go-erofs"
-	"golang.org/x/sys/unix"
 )
 
 // NewOverlayFS returns a new fs.FS that overlays the provided layers.
@@ -236,48 +231,4 @@ func (d *overlayDir) ReadDir(n int) ([]fs.DirEntry, error) {
 	res := d.entries[d.offset:end]
 	d.offset = end
 	return res, nil
-}
-
-func isWhiteout(fi fs.FileInfo) bool {
-	if (fi.Mode() & fs.ModeCharDevice) == 0 {
-		return false
-	}
-
-	// Check for regular syscall.Stat_t (from os.File)
-	if sys, ok := fi.Sys().(*syscall.Stat_t); ok {
-		return sys.Rdev == 0
-	}
-
-	// Check for EROFS Stat
-	if estatfi, ok := fi.Sys().(*erofs.Stat); ok {
-		return estatfi.Rdev == 0
-	}
-
-	return false
-}
-
-func isOpaque(f fs.File) bool {
-	// Attempt to get underlying *os.File
-	if osf, ok := f.(*os.File); ok {
-		dest := make([]byte, 1)
-		sz, err := unix.Fgetxattr(int(osf.Fd()), "trusted.overlay.opaque", dest)
-		if err != nil {
-			return false
-		}
-		return sz == 1 && dest[0] == 'y'
-	}
-
-	// Check for EROFS file by getting Stat and checking xattrs
-	fi, err := f.Stat()
-	if err != nil {
-		return false
-	}
-
-	if estatfi, ok := fi.Sys().(*erofs.Stat); ok {
-		if xattr, ok := estatfi.Xattrs["trusted.overlay.opaque"]; ok {
-			return xattr == "y"
-		}
-	}
-
-	return false
 }
