+++ internal/fsview/overlay_other.go
@@ -1,0 +1,54 @@
+//go:build !linux
+
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package fsview
+
+import (
+	"io/fs"
+
+	"github.com/erofs/go-erofs"
+)
+
+func isOpaque(f fs.File) bool {
+	// Check for EROFS file by getting Stat and checking xattrs
+	fi, err := f.Stat()
+	if err != nil {
+		return false
+	}
+
+	if estatfi, ok := fi.Sys().(*erofs.Stat); ok {
+		if xattr, ok := estatfi.Xattrs["trusted.overlay.opaque"]; ok {
+			return xattr == "y"
+		}
+	}
+
+	return false
+}
+
+func isWhiteout(fi fs.FileInfo) bool {
+	if (fi.Mode() & fs.ModeCharDevice) == 0 {
+		return false
+	}
+
+	// Check for EROFS Stat
+	if estatfi, ok := fi.Sys().(*erofs.Stat); ok {
+		return estatfi.Rdev == 0
+	}
+
+	return false
+}
